requirements.txt
absl-py==0.4.0
astor==0.7.1
certifi==2018.11.29
cffi==1.12.1
Click==7.0
cmake==3.12.0
dlib==19.15.0
face-recognition==1.2.3
face-recognition-models==0.3.0
ffmpy==0.2.2
ffmpeg
gast==0.2.0
grpcio==1.14.1
h5py==2.10.0
icc-rt==2019.0
intel-numpy==1.15.1
intel-openmp==2019.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
Markdown==2.6.11
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1.1
munch==2.3.2
numpy==1.17.4
olefile==0.46
opencv-python==3.4.1.15
pathlib==1.0.1
Pillow==7.2.0
pretrainedmodels==0.7.4
protobuf==3.6.1
pycparser==2.19
PyYAML==5.1.1
scandir==1.7
scipy==1.3.0
six==1.12.0
tbb==2019.0
tbb4py==2019.0
termcolor==1.1.0
tensorflow-gpu==1.15.0
torch==1.0.1.post2
torchvision==0.2.1
tqdm==4.25.0
youtube-dl